refactor: Query Packaging
