+++ src/main/java/com/commerce/backendserver/product/application/ProductFindService.java
@@ -1,7 +1,6 @@
 package com.commerce.backendserver.product.application;
 
-import com.commerce.backendserver.product.domain.Product;
-import com.commerce.backendserver.product.domain.ProductQueryRepository;
+import com.commerce.backendserver.product.domain.persistence.ProductQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
