--- src/main/java/com/commerce/backendserver/product/domain/ProductQueryRepository.java
@@ -1,10 +1,0 @@
-package com.commerce.backendserver.product.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-@Repository
-@Transactional(readOnly = true)
-public interface ProductQueryRepository extends JpaRepository<Product, Long> {
-}
