+++ src/main/java/com/commerce/backendserver/product/domain/infra/query/ProductQueryDslRepositoryImpl.java
@@ -1,0 +1,4 @@
+package com.commerce.backendserver.product.domain.infra.query;
+
+public class ProductQueryDslRepositoryImpl implements ProductQueryDslRepository{
+}
