+++ src/main/java/com/commerce/backendserver/product/domain/persistence/ProductQueryRepository.java
@@ -1,0 +1,8 @@
+package com.commerce.backendserver.product.domain.persistence;
+
+import com.commerce.backendserver.product.domain.infra.query.ProductQueryDslRepository;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional(readOnly = true)
+public interface ProductQueryRepository extends ProductJpaRepository, ProductQueryDslRepository {
+}
